feat(purchase): standalone Goods Receipts, over-receipt allowance, decimal qty
Three operational blockers for bulk items (semen curah/jumbo) received by
truck-scale weight:

- Goods Receipt no longer requires a Purchase Order — a Direct Receipt mode
  lets Supplier/Warehouse/items be entered manually (GoodsReceiptService::
  createDirect()), with nullable purchase_order_id/purchase_order_item_id
  and "Direct Receipt" fallbacks everywhere a PO link was assumed.
- Per-Item "Allow Over-Receipt" flag lets a line exceed the PO's outstanding
  qty (assertWithinOutstanding()) instead of hard-blocking it.
- Purchase quantities (PO/GR/Invoice/Return + the shared stock ledger) move
  from integer to decimal(15,2), end to end: DB columns, model casts,
  StockLedgerService's shared type hints, request validation, and every
  frontend qty input/schema in the Purchase module.

Widening qty from an integer to a decimal cast changes its JSON wire format
from a number to a string (same as rate/amount already are) — fixed every
raw `sum + qty`-style reduce and strict `=== 0`/`(int)` truncation this
exposed across Purchase, Inventory, and the Dashboard.

+++ backend/laravel-api/app/Http/Resources/PurchaseInvoiceItemResource.php
@@ -24,9 +24,9 @@ class PurchaseInvoiceItemResource extends JsonResource
             'rate' => $this->rate,
             'qty' => $this->qty,
             'amount' => $this->amount,
-            'returned_qty' => (int) $returnedTotals->qty,
+            'returned_qty' => (float) $returnedTotals->qty,
             'returned_amount' => (float) $returnedTotals->amount,
-            'returnable_qty' => (int) $this->qty - (int) $returnedTotals->qty,
+            'returnable_qty' => (float) $this->qty - (float) $returnedTotals->qty,
             'returnable_amount' => (float) $this->amount - (float) $returnedTotals->amount,
         ];
     }
